Move IntRect into Geometry namespace
And move IntRectExtensions to Geometry.Extensions namespace
