+++ Assets/Scripts/Geometry/Extensions/IntRectExtensions.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace PAC.DataStructures.Extensions
+namespace PAC.Geometry.Extensions
 {
     /// <summary>
     /// Extension methods for <see cref="IntRect"/>.
