+++ Assets/Scripts/Geometry/Extensions/IntRectExtensions.cs.meta
