+++ Assets/Scripts/Geometry/Shapes/Diamond.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
-using PAC.DataStructures;
 using PAC.Extensions;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes.Interfaces;
