+++ Assets/Scripts/Geometry/Shapes/Ellipse.cs
@@ -2,8 +2,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
-
-using PAC.DataStructures;
 using PAC.Maths;
 using PAC.Geometry.Shapes.Interfaces;
