+++ Assets/Scripts/Geometry/Shapes/IsometricRectangle.cs
@@ -3,7 +3,6 @@ using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 
-using PAC.DataStructures;
 using PAC.Exceptions;
 using PAC.Extensions;
 using PAC.Geometry.Axes;
