+++ Assets/Scripts/Image Editing/ImageEdit.cs
@@ -1,7 +1,5 @@
 using System;
 
-using PAC.DataStructures;
-using PAC.DataStructures.Extensions;
 using PAC.Exceptions;
 using PAC.Extensions;
 using PAC.Geometry;
