+++ Assets/Scripts/Image Editing/Outline.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 
-using PAC.DataStructures.Extensions;
 using PAC.Extensions;
 using PAC.Geometry;
 using PAC.Geometry.Extensions;
