+++ Assets/Scripts/Image Editing/Texture2DCreator.cs
@@ -1,8 +1,8 @@
 using System;
 
 using PAC.Colour;
-using PAC.DataStructures;
 using PAC.Extensions;
+using PAC.Geometry;
 using PAC.Patterns;
 using PAC.Patterns.Extensions;
