+++ Assets/Tests/Geometry/IntRect_Tests.cs
@@ -4,7 +4,6 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Extensions;
 using PAC.Geometry;
