+++ Assets/Tests/Geometry/Shapes/Default Tests/IShape_DefaultTests.cs
@@ -4,7 +4,6 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Exceptions;
 using PAC.Geometry;
 using PAC.Geometry.Shapes.Interfaces;
