+++ Assets/Tests/Geometry/Shapes/IsometricRectangle_Tests.cs
@@ -1,6 +1,5 @@
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes;
