+++ Assets/Tests/Geometry/Shapes/Path_Tests.cs
@@ -4,7 +4,6 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Extensions;
 using PAC.Geometry;
 using PAC.Geometry.Shapes;
