+++ Assets/Tests/Geometry/Shapes/TestUtils/ShapeAssert.cs
@@ -3,7 +3,6 @@ using System.Linq;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
 using PAC.Geometry.Shapes.Interfaces;
