+++ Assets/Tests/Patterns/GradientRadial_Tests.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 
 using NUnit.Framework;
 
-using PAC.DataStructures;
 using PAC.Geometry;
 using PAC.Geometry.Axes;
